+++ src/ts/src/public/cycle/CycleDeleteRequest.ts
@@ -1,7 +1,0 @@
-export interface CycleDeleteRequest {
-  cycleId: string;
-}
-
-export interface CycleDeleteResponse {
-  /* Empty */
-}