+++ src/ts/src/public/cycle/CycleGetRequest.ts
@@ -1,5 +1,0 @@
-import { CycleItemDto } from "./dto";
-
-export interface CycleGetResponse {
-  item: CycleItemDto;
-}