+++ src/ts/src/public/personal-goals/PersonalGoalsGetRequest.ts
@@ -1,0 +1,8 @@
+import { PersonalGoalsDto } from "./dto";
+
+export interface PersonalGoalsGetRequest {
+  /* Empty on purpose */
+}
+export interface PersonalGoalsGetResponse {
+  item: PersonalGoalsDto;
+}