+++ src/ts/src/public/personal-goals/dto.ts
@@ -1,0 +1,17 @@
+export interface PersonalGoalsDto {
+  accountId: string;
+  goals: PersonalGoal[];
+  /** The order in which tasks (including goals) should be displayed */
+  order: PersonalGoal[];
+}
+
+export interface PersonalGoalsUpdateValues {
+  goals?: PersonalGoal[];
+  /** The order in which tasks (including goals) should be displayed */
+  order?: PersonalGoal[];
+}
+
+export interface PersonalGoal {
+  workspaceId: string;
+  taskId: string;
+}