+++ src/ts/src/public/personal-goals/events.ts
@@ -1,0 +1,5 @@
+import { PersonalGoalsDto } from "./dto";
+
+export interface PersonalGoalsSetEvent {
+  item: PersonalGoalsDto;
+}