+++ src/ts/src/public/personal-goals/index.ts
@@ -1,0 +1,5 @@
+export * from "./api";
+export * from "./dto";
+export * from "./events";
+export * from "./PersonalGoalsGetRequest";
+export * from "./PersonalGoalsUpdateRequest";